Create Reverse_Palindrome.java
Check whether a given number is a palindrome or not. Print the number, if it is a palindrome, otherwise add the number to its reverse and perform this operation until it becomes a palindrome.
Input format: An integer in a single line
Output format: An integer in a single line. 
Sample Input: 195
Sample Output: 9339